requirement.txt
streamlit
tensorflow
numpy
pillow
googletrans==4.0.0-rc1